+++ morphir/concepts/src/org/finos/morphir/concepts/Distro.scala
@@ -1,0 +1,8 @@
+package org.finos
+package morphir.concepts
+
+import morphir.core.types.Versioning.MorphirVersion
+
+enum Distro:
+  case Library
+  case Bundle(morphirVersion: MorphirVersion)
