+++ morphir/concepts/src/org/finos/morphir/concepts/workspace/Workspace.scala
@@ -1,0 +1,9 @@
+package org.finos
+package morphir
+package concepts
+package workspace
+
+import java.nio.file.Path
+
+final case class Workspace(path: Path)
+object Workspace
