+++ morphir/concepts/src/org/finos/morphir/implicits/Readers.scala
@@ -1,0 +1,12 @@
+package org.finos
+package morphir
+package implicits
+
+import morphir.core.*
+import org.finos.morphir.core.types.Name
+trait Readers extends morphir.core.Types:
+  given NameReader: Reader[Name] = new SimpleReader[Name] {
+    override def expectedMsg: String = "expected Name"
+
+    override def visitName(value: CharSequence, index: Int): Name = Name(value)
+  }
