+++ morphir/core/src/org/finos/morphir/core/package.scala
@@ -1,5 +1,7 @@
 package org.finos.morphir
 
+import org.finos.morphir.core.types.Name
+
 package object core:
   export org.finos.morphir.core.internal.InternalTypes.*
   export org.finos.morphir.core.internal.Visitor
