+++ morphir/core/src/org/finos/morphir/core/types/Versioning.scala
@@ -1,0 +1,7 @@
+package org.finos.morphir.core.types
+
+object Versioning:
+  opaque type MorphirVersion = String
+
+  object MorphirVersion:
+    def apply(version: String): MorphirVersion = version
