--- morphir/toolkit/vfile/src/org/finos/morphir/toolkit/vfile/package.scala
@@ -1,6 +1,0 @@
-package org.finos.morphir.toolkit
-
-package object vfile {
-  type VFileRepr[+A] = VFile.Repr[A]
-  val VFileRepr = VFile.Repr
-}
