+++ morphir/vfile/src/org/finos/morphir/toolkit/vfile/VMessage.scala
