+++ morphir/vfile/src/org/finos/morphir/toolkit/vfile/package.scala
@@ -1,0 +1,3 @@
+package org.finos.morphir.toolkit
+
+package object vfile {}
